+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+...
+
+## [0.10.2] - 2019-09-05
+
+### Added
+
+- Foreign Key constraints can now be added to tables using new method `DiscoveredTable.AddForeignKey`
+
+### Fixed
+
+- Fixed bug in MySql where `DiscoveredTable.DiscoverRelationships(...)` could throw an ArgumentException ("same key has already been added [...]") in some circumstances
+ 
 ## [0.10.1] - 2019-09-05
 
 ### Fixed
@@ -100,7 +112,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Drop table to work correctly with Views
 - Exists now works correctly for Views (previously it would return true if there was no view but a table with the same name)
 
-[Unreleased]: https://github.com/HicServices/FAnsiSql/compare/0.10.1...develop
+[Unreleased]: https://github.com/HicServices/FAnsiSql/compare/0.10.2...develop
+[0.10.2]: https://github.com/HicServices/FAnsiSql/compare/0.10.1...0.10.2
 [0.10.1]: https://github.com/HicServices/FAnsiSql/compare/0.10.0...0.10.1
 [0.10.0]: https://github.com/HicServices/FAnsiSql/compare/0.9.8...0.10.0
 [0.9.8]: https://github.com/HicServices/FAnsiSql/compare/0.9.7...0.9.8